Switch to safer method of cloning request bodies
The multi-pipe clone method was hitting an issue where the read end of
one of the pipes was getting closed. This caused a panic on the write
end. Switch to the safer method of copying the entire request into
memory first, and then creating two io.Readers. This will use more
memory, but so be it.